improved memory management utilities
OString values are strings created and managed by OCaml (although they
live in the C heap).

Opaque values are abstract values that are managed by OCaml (and live in
the OCaml heap).

Both kind of values will not be garbage collected, unless explicitly
freed with a polymorphic `bap_free` function.