prepare project to support mutliple libraries
The plan is to build different libraries while reusing the same
harness. Each new library will be added in a separate folder. One
library can be built at once.

This commit also moves all configuration paraphernalia into the
configure.ac, so that the invocation of Makefile doesn't depend
on the evironment or a user.