fixes the default disassembler selection
After we added a new disassembler (Ghidra) in BAP 2.4.0 the method for
selecting a default disassembler (which is just pick the first
available), no longer makes sense (as Ghidra requires different
inputs).

The solution is simply to hardcode llvm as the default
disassembler. The Ghidra disassembler could still be used via the
`disasm_basic_create_expert` function.